Structure repo in bin and scampi. Made it pip installable via setup.py. Changed imports to scampi (small letters). Added function data_handling.py to write an array to disk, but it needs depugging.
