+++ .gitignore
@@ -1,0 +1,8 @@
+# Ignore compiled Python bytecode files
+*.pyc
+
+# Ignore Python cache directories
+*__pycache__/
+
+# Python egg metadata, regenerated from source files by setuptools.
+/*.egg-info
