+++ bin/best_fit_alpha_tau.py
@@ -1,4 +1,4 @@
-#!/bin/python
+#! /usr/bin/env python3
 
 import argparse
 import emcee
