+++ bin/define_alpha_burnin.py
@@ -1,6 +1,5 @@
 #! /usr/bin/env python3
 
-import numpy as np
 import argparse
 import pandas
 
@@ -26,4 +25,4 @@ if __name__ == '__main__':
         pass
     # Write to new file.
     psrinfo.to_csv(args.run_log)
-    
+
