+++ scampi/model_functions.py
