+++ test_example/examplerun.sh
@@ -33,16 +33,16 @@ mkdir -p output/results/plots
 # RUN THE SCRIPTS
 ##################
 echo "Fitting for tau..."
-python $MYCODE/SCAMP_I/run_scatter_mcmc.py -f ${configfile} -o logs/${basename}_inputlog.csv -w output/chains -t ${taumcruntime} -nw ${tau_nwalkers} -m ${model} --showprogress
-python $MYCODE/SCAMP_I/check_tau_chains.py -i logs/${basename}_inputlog.csv -w output/diagnostic_plots
+run_scatter_mcmc.py -f ${configfile} -o logs/${basename}_inputlog.csv -w output/chains -t ${taumcruntime} -nw ${tau_nwalkers} -m ${model} --showprogress
+check_tau_chains.py -i logs/${basename}_inputlog.csv -w output/diagnostic_plots
 echo "Now check the chain diagnostic plots saved in the output/diagnostic_plots/ folder and decide whether they have passed or failed and what burnin to apply."
-python $MYCODE/SCAMP_I/define_tau_passfail_burnin.py -i logs/${basename}_inputlog.csv -o ${basename}_outputlog.csv -w logs
-python $MYCODE/SCAMP_I/parameter_extraction.py -i logs/${basename}_outputlog.csv -o ${basename}_output.csv -w output/results
+define_tau_passfail_burnin.py -i logs/${basename}_inputlog.csv -o ${basename}_outputlog.csv -w logs
+parameter_extraction.py -i logs/${basename}_outputlog.csv -o ${basename}_output.csv -w output/results
 echo "Fitting for alpha..."
-python $MYCODE/SCAMP_I/alpha_mcmc.py -i output/results/${basename}_output.csv -o ${basename}_output.csv --sampleswritedir output/chains --resultswritedir output/results -t ${alphamcruntime} -nw ${alpha_nwalkers} --showprogress
-python $MYCODE/SCAMP_I/check_alpha_chains.py -i output/results/${basename}_output.csv -w output/diagnostic_plots
+alpha_mcmc.py -i output/results/${basename}_output.csv -o ${basename}_output.csv --sampleswritedir output/chains --resultswritedir output/results -t ${alphamcruntime} -nw ${alpha_nwalkers} --showprogress
+check_alpha_chains.py -i output/results/${basename}_output.csv -w output/diagnostic_plots
 echo "Now check the alpha chain plots and decide what burnin fraction you want."
-python $MYCODE/SCAMP_I/define_alpha_burnin.py -i output/results/${basename}_output.csv
-python $MYCODE/SCAMP_I/best_fit_alpha_tau.py -i output/results/${basename}_output.csv
-python $MYCODE/SCAMP_I/basic_plotting.py -f output/results/${basename}_output.csv -w output/results/plots
+define_alpha_burnin.py -i output/results/${basename}_output.csv
+best_fit_alpha_tau.py -i output/results/${basename}_output.csv
+basic_plotting.py -f output/results/${basename}_output.csv -w output/results/plots
 echo "Results and plots saved to directories."
